split utils into separate files, continue simplification, added choice between tcp and udp when using gstreamer
