+++ integration_tests/Cargo.toml
@@ -9,9 +9,7 @@ license = "BUSL-1.1"
 [features]
 default = ["web_renderer"]
 update_snapshots = []
-decklink = [
-    "live_compositor/decklink",
-]
+decklink = ["live_compositor/decklink"]
 web_renderer = [
     "dep:compositor_chromium",
     "compositor_render/web_renderer",
@@ -45,3 +43,5 @@ compositor_chromium = { path = "../compositor_chromium", optional = true }
 tokio-tungstenite = "0.21.0"
 wgpu = { workspace = true }
 image = { workspace = true }
+strum = "0.26"
+strum_macros = "0.26"
