+++ integration_tests/examples/aac_rtp_input.rs
@@ -8,8 +8,9 @@ use std::{
     time::Duration,
 };
 
-use integration_tests::examples::{
-    self, download_file, ff_stream_audio, ff_stream_video, start_ffplay, start_websocket_thread,
+use integration_tests::{
+    ffmpeg_utils::{start_ffmpeg_receive, start_ffmpeg_send_audio, start_ffmpeg_send_video},
+    utils::{self, download_file, start_websocket_thread},
 };
 
 const BUNNY_FILE_URL: &str =
@@ -44,14 +45,14 @@ fn main() {
 
 fn start_example_client_code() -> Result<()> {
     info!("[example] Start listening on output port.");
-    start_ffplay(IP, Some(OUTPUT_VIDEO_PORT), Some(OUTPUT_AUDIO_PORT))?;
+    start_ffmpeg_receive(Some(OUTPUT_VIDEO_PORT), Some(OUTPUT_AUDIO_PORT))?;
     start_websocket_thread();
 
     info!("[example] Download sample.");
     let bunny_path = download_file(BUNNY_FILE_URL, BUNNY_FILE_PATH)?;
 
     info!("[example] Send register input request.");
-    examples::post(
+    utils::post(
         "input/input_1/register",
         &json!({
             "type": "rtp_stream",
@@ -63,7 +64,7 @@ fn start_example_client_code() -> Result<()> {
     )?;
 
     info!("[example] Send register input request.");
-    examples::post(
+    utils::post(
         "input/input_2/register",
         &json!({
             "type": "rtp_stream",
@@ -81,7 +82,7 @@ fn start_example_client_code() -> Result<()> {
     )?;
 
     info!("[example] Send register output request.");
-    examples::post(
+    utils::post(
         "output/output_1/register",
         &json!({
             "type": "rtp_stream",
@@ -108,7 +109,7 @@ fn start_example_client_code() -> Result<()> {
     )?;
 
     info!("[example] Send register output request.");
-    examples::post(
+    utils::post(
         "output/output_2/register",
         &json!({
             "type": "rtp_stream",
@@ -131,10 +132,10 @@ fn start_example_client_code() -> Result<()> {
     std::thread::sleep(Duration::from_millis(500));
 
     info!("[example] Start pipeline");
-    examples::post("start", &json!({}))?;
+    utils::post("start", &json!({}))?;
 
-    ff_stream_video(IP, INPUT_1_PORT, bunny_path.clone())?;
-    ff_stream_audio(IP, INPUT_2_PORT, bunny_path, "aac")?;
+    start_ffmpeg_send_video(IP, INPUT_1_PORT, bunny_path.clone())?;
+    start_ffmpeg_send_audio(IP, INPUT_2_PORT, bunny_path, "aac")?;
 
     Ok(())
 }
