+++ integration_tests/examples/transition.rs
@@ -4,8 +4,9 @@ use log::{error, info};
 use serde_json::json;
 use std::{thread, time::Duration};
 
-use integration_tests::examples::{
-    self, start_ffplay, start_websocket_thread, ff_stream_testsrc,
+use integration_tests::{
+    ffmpeg_utils::{start_ffmpeg_receive, start_ffmpeg_send_testsrc},
+    utils::{self, start_websocket_thread},
 };
 
 const VIDEO_RESOLUTION: Resolution = Resolution {
@@ -31,11 +32,11 @@ fn main() {
 
 fn start_example_client_code() -> Result<()> {
     info!("[example] Start listening on output port.");
-    start_ffplay(IP, Some(OUTPUT_PORT), None)?;
+    start_ffmpeg_receive(Some(OUTPUT_PORT), None)?;
     start_websocket_thread();
 
     info!("[example] Send register input request.");
-    examples::post(
+    utils::post(
         "input/input_1/register",
         &json!({
             "type": "rtp_stream",
@@ -48,7 +49,7 @@ fn start_example_client_code() -> Result<()> {
 
     let shader_source = include_str!("./silly.wgsl");
     info!("[example] Register shader transform");
-    examples::post(
+    utils::post(
         "shader/example_shader/register",
         &json!({
             "source": shader_source,
@@ -56,7 +57,7 @@ fn start_example_client_code() -> Result<()> {
     )?;
 
     info!("[example] Register static image");
-    examples::post(
+    utils::post(
         "image/example_image/register",
         &json!({
             "asset_type": "gif",
@@ -157,7 +158,7 @@ fn start_example_client_code() -> Result<()> {
     });
 
     info!("[example] Send register output request.");
-    examples::post(
+    utils::post(
         "output/output_1/register",
         &json!({
             "type": "rtp_stream",
@@ -180,15 +181,15 @@ fn start_example_client_code() -> Result<()> {
     )?;
 
     info!("[example] Start pipeline");
-    examples::post("start", &json!({}))?;
+    utils::post("start", &json!({}))?;
 
     info!("[example] Start input stream");
-    ff_stream_testsrc(IP, INPUT_PORT, VIDEO_RESOLUTION)?;
+    start_ffmpeg_send_testsrc(IP, INPUT_PORT, VIDEO_RESOLUTION)?;
 
     thread::sleep(Duration::from_secs(5));
 
     info!("[example] Update output");
-    examples::post(
+    utils::post(
         "output/output_1/update",
         &json!({
             "video": {
@@ -200,7 +201,7 @@ fn start_example_client_code() -> Result<()> {
     thread::sleep(Duration::from_secs(2));
 
     info!("[example] Update output");
-    examples::post(
+    utils::post(
         "output/output_1/update",
         &json!({
             "video": {
